Add initial version of cmake related files + some other bits :-p
********************************************************************************
                          I M P O R T A N T
********************************************************************************

This commit does not represent anything 'finished' and is very much a work in
progress.  It's okay to try it out, but not to whine, patch, or even give much
in the way of feedback.  The sole reason this is being committed is in order
to more easily keep things from getting too out dated as compared with current
trunk.

This said, the current status is considered to be:

All builds
==========
Static building is at the moment totally untested.


Linux & similar systems
=======================
Everything is about 99% completed.  Known exceptions are the missing .pc files
and the fact that the python modules are named incorrectly and get installed in
the wrong place :)


Windows / MSVC++
================
Again, things are largely complete.  Everything will build ok.  There's probably
some work to do for detecting certain optional libs, and also the python modules
have the incorrect file extension.


Apple Mac OS X / Xcode
======================
Things are moving in a slightly different direction here, since I'm moving away
from eveything being a framework - which was total overkill - to mostly using
the more basic .dylib files.  Due to this, a new dependency pack is required but
is not available at this time (you can rework the existing dependency binaries to
be non framework based if you know what you are doing).  In addition, while
everything will basically build, there is still some work to do: for example the
sample app bundles do not yet have the dependency files copied / symlinked into
the appropriate places (though if you do it manually, you can get them to run).

********************************************************************************